requirements.txt
flask
azure-identity
azure-keyvault-secrets
psycopg2-binary